Simple dynamic memory allocator and basic printf implementation

+++ libc/memory/memory.h
@@ -2,4 +2,4 @@
 #include "types.h"
 
 void memcpy(void* dest, const void* src, uint64_t n);
-void* memset(void* dest, int val, uint64_t n);
+void* memset(void* dest, uint8_t val, uint64_t n);
