+++ libc/memory/memset.c
@@ -1,6 +1,6 @@
 #include "memory.h"
 
-void* memset(void* dest, int val, uint64_t n) {
+void* memset(void* dest, uint8_t val, uint64_t n) {
     for (uint64_t i = 0; i < n; i++) {
         ((uint8_t*)dest)[i] = (uint8_t)val;
     }
